+++ NEWS.md
@@ -98,6 +98,7 @@ Red Panda C++ Version 2.27
   - enhancement: Adjust function tip pos to prevent it from run outside the right window edge.
   - enhancement: Open ".def" (Module definition file) file in editor when double click it in the project view.
   - enhancement: When a dll project has .def file, use it when generating the dll file.  
+  - fix: "project name".exe.manifest is auto removed when build the project.
     
 Red Panda C++ Version 2.26
   - enhancement: Code suggestion for embedded std::vectors.